Currency converter component added and functionnal, there is some issue with styling so it need to be reviewed, the branch currency will be merged to master for the review since the component is functionnal and has shown no issue
